Bug #306020 --> hard code some spots in the UI to say Thunderbird Beta 1. or version 1.5 Beta 1
Beta one is going to have an application version of 1.4 for software update purposes. But we don't want users thinking
they are using some mythical 1.4 version. So make the about dialog say 1.5 beta 1. We'll change these strings to Beta 2
and then remove them all together when the time is right.

sr=bsmedberg
a=me