Windows CE ONLY
1. Use PostMessages as we do not care about the immediate result nor wish to
wait.

2. Remove gSoftKeyMenuBar as we want all toplevel, dialog, and popup windows to
have a softkey menu bar.

3. Reverting the SWP_NOACTIVATE flag to SetWindowPos.  It really seams that we
can't click on the popup's without this.  Because of this change, I also
reverting the check in DealWithPopups for the WM_ACTIVATE flag.

4. When handling the gesture in left mouse down, we always want to return zero.
Enabling WM_IME_ on windows ce.  The APIs are present, we just were ifdef'ing
out the messaging handling.

sr=dveditz, a=mconnor