fix for bug #364599: some newly created profile files are write protected.
For Firefox 2.0.0.1-de, the default bookmarks.html, localstore.rdf, and mimeTypes.rdf (as well as search.rdf, but that is unused) files had the wrong permissions, which caused several problems.

Going forward, bsmedberg has made it so the build system will fix this problem if it happens again. But for those 2.0.0.1-de users who were affected, this change will fix the permissions on these files.

r=bsmedberg, dveditz, a=dveditz